requirements.txt
elasticsearch
coverage
pyyaml